Issue #378 Ensure public API of Unicorn.FontTools.OpenType is CLS-compliant even when the types implied by the OpenType binary format are not

+++ Unicorn.FontTools.OpenType.Tests.Unit/Mocks/MockCharacterMapping.cs
@@ -14,17 +14,17 @@ namespace Unicorn.FontTools.OpenType.Tests.Unit.Mocks
             throw new NotImplementedException(TestResources.Mocks_MockCharacterMapping_NotImplementedError);
         }
 
-        public override ushort MapCodePoint(byte codePoint)
+        public override int MapCodePoint(byte codePoint)
         {
             throw new NotImplementedException(TestResources.Mocks_MockCharacterMapping_NotImplementedError);
         }
 
-        public override ushort MapCodePoint(ushort codePoint)
+        public override int MapCodePoint(int codePoint)
         {
             throw new NotImplementedException(TestResources.Mocks_MockCharacterMapping_NotImplementedError);
         }
 
-        public override ushort MapCodePoint(uint codePoint)
+        public override int MapCodePoint(long codePoint)
         {
             throw new NotImplementedException(TestResources.Mocks_MockCharacterMapping_NotImplementedError);
         }
